project/requirements.txt
asgiref==3.8.1
Django==4.2.14
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
importlib_metadata==8.2.0
Markdown==3.6
mysqlclient==2.2.4
PyJWT==2.8.0
sqlparse==0.5.1
typing_extensions==4.12.2
zipp==3.19.2
